requirements.txt
qbittorrent-api
pyyaml
flask
apscheduler
serverchan-sdk
gunicorn
requests